pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "unomd"
version = "0.1.0"
description = "A package for molecular dynamics simulations"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Ingrid Barbosa-Farias", email = "ingrid@simatomic.com"},
    {name = "Omar Arias Gaguancela", email = "omar.arias@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openmm>=7.7.0",
    "mdanalysis>=2.4.0",
    "mdtraj>=1.9.7",
    "numpy>=1.23.5",
    "scipy>=1.9.3",
    "pandas>=1.5.3",
    "tqdm>=4.64.1",
    "pyyaml>=6.0",
    "h5py>=3.7.0",
    "plotly>=5.18.0",
    "py3dmol>=0.1.0"
]

[project.optional-dependencies]
dev = [
    "black>=22.12",
    "flake8>=6.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "jupyter",
    "ipywidgets>=8.0",
]

test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10.0",
    "pytest-timeout>=2.1.0",
]

[project.urls]
Repository = ""  # Add your repository URL here

[tool.hatch.build.targets.wheel]
packages = ["src/unomd"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=unomd --cov-report=term-missing"
timeout = 300 